+++ node/test/benchmark/lib/basename.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const testUid = require('../../lib/test-uid');
+
+module.exports = `test-benchmark-${testUid}`;