+++ node/test/benchmark/lib/cleanup.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const cleanup = require('../../lib/cleanup');
+const basename = require('./basename');
+
+module.exports = async (options = {}) =>
+  cleanup(basename, { extraLayerNames: ['node-internal', 'python-internal'], ...options });