+++ node/test/lib/get-process-function.js
@@ -25,7 +25,7 @@ module.exports = async (basename, coreConfig, options) => {
   const baseLambdaConfiguration = _.merge(
     {
       Role: coreConfig.roleArn,
-      MemorySize: 1024,
+      MemorySize: Number(options.lambdaMemorySize || 1024),
       Layers: [coreConfig.layerInternalArn],
       Environment: {
         Variables: {
@@ -42,7 +42,7 @@ module.exports = async (basename, coreConfig, options) => {
   const { TracePayload, LogPayload, DevModePayload } = options;
   ensurePlainFunction(TracePayload.encode);
 
-  if (!baseLambdaConfiguration.Code) {
+  if (!baseLambdaConfiguration.Code && options.fixturesDirname) {
     baseLambdaConfiguration.Code = { ZipFile: await resolveDirZipBuffer(options.fixturesDirname) };
   }